refactor: extract council discussion logic into server/councils/discussion.ts
Move all council business logic (launch, discussion, review, synthesis,
abort, chat, auto-advance, session waiting) from routes/councils.ts into
a dedicated server/councils/discussion.ts module. The route file now
contains only thin HTTP handlers that delegate to the extracted functions.
Re-exports maintain backward compatibility for all existing consumers.

Closes #278
